requirements.txt
Django==1.6
South==0.8.3
django-registration==1.0
django-debug-toolbar==1.0
requests==2.2.0
tmdbsimple==0.3.0
psycopg2==2.5.2
gunicorn==18.0
dj-database-url==0.2.2
dj-static==0.0.5
boto==2.23.0
django-storages==1.1.8
celery==3.1.8
django-celery==3.1.1
newrelic==2.10.1.9
